+++ sadmin/controller/catalog/team.php
@@ -23,8 +23,8 @@ class ControllerCatalogTeam extends PT_Controller
 
         $this->load->model('catalog/team');
 
-        if (($this->request->server['REQUEST_METHOD'] == 'POST')) {
-           
+        if (($this->request->server['REQUEST_METHOD'] == 'POST') && $this->validateForm()) {
+
             $this->model_catalog_team->addTeam($this->request->post);
 
             $this->session->data['success'] = $this->language->get('text_success');
@@ -43,7 +43,7 @@ class ControllerCatalogTeam extends PT_Controller
 
         $this->load->model('catalog/team');
 
-        if (($this->request->server['REQUEST_METHOD'] == 'POST')) {
+        if (($this->request->server['REQUEST_METHOD'] == 'POST') && $this->validateForm()) {
             $this->model_catalog_team->editTeam($this->request->get['team_id'], $this->request->post);
 
             $this->session->data['success'] = $this->language->get('text_success');
@@ -62,7 +62,7 @@ class ControllerCatalogTeam extends PT_Controller
 
         $this->load->model('catalog/team');
 
-        if (isset($this->request->post['selected'])) {
+        if (isset($this->request->post['selected']) && $this->validateDelete()) {
             foreach ($this->request->post['selected'] as $team_id) {
                 $this->model_catalog_team->deleteTeam($team_id);
             }
@@ -121,6 +121,7 @@ class ControllerCatalogTeam extends PT_Controller
                 'name'          => $result['name'],
                 'designation'   => $result['designation'],
                 'sort_order'    => $result['sort_order'],
+                'status'        => ($result['status'] ? $this->language->get('text_enabled') : $this->language->get('text_disabled')),
                 'date_added'    => date($this->language->get('date_format_short'), strtotime($result['date_added'])),
                 'edit'          => $this->url->link('catalog/team/edit', 'user_token=' . $this->session->data['user_token'] . '&team_id=' . $result['team_id'])
             );
@@ -159,7 +160,7 @@ class ControllerCatalogTeam extends PT_Controller
         $this->document->addScript("view/dist/plugins/ckeditor/ckeditor.js");
         $this->document->addScript("view/dist/plugins/ckeditor/adapters/jquery.js");
         $this->document->addScript("view/dist/plugins/iCheck/icheck.min.js");
-        
+
         $data['text_form'] = !isset($this->request->get['team_id']) ? $this->language->get('text_add') : $this->language->get('text_edit');
 
         if (isset($this->error['warning'])) {
@@ -219,7 +220,7 @@ class ControllerCatalogTeam extends PT_Controller
         } else {
             $data['name'] = '';
         }
-        
+
         if (isset($this->request->post['designation'])) {
             $data['designation'] = $this->request->post['designation'];
         } elseif (!empty($team_info)) {
@@ -227,7 +228,7 @@ class ControllerCatalogTeam extends PT_Controller
         } else {
             $data['designation'] = '';
         }
-        
+
         if (isset($this->request->post['description'])) {
             $data['description'] = $this->request->post['description'];
         } elseif (!empty($team_info)) {
@@ -243,7 +244,7 @@ class ControllerCatalogTeam extends PT_Controller
         } else {
             $data['sort_order'] = 0;
         }
-        
+
         if (isset($this->request->post['image'])) {
             $data['image'] = $this->request->post['image'];
         } elseif (!empty($team_info)) {
@@ -267,7 +268,7 @@ class ControllerCatalogTeam extends PT_Controller
         } elseif (!empty($team_info)) {
             $data['status'] = $team_info['status'];
         } else {
-            $data['status'] = 0;
+            $data['status'] = true;
         }
 
         $data['header'] = $this->load->controller('common/header');
@@ -283,34 +284,16 @@ class ControllerCatalogTeam extends PT_Controller
             $this->error['warning'] = $this->language->get('error_permission');
         }
 
-        if ((utf8_strlen($this->request->post['name']) > 96) || !filter_var($this->request->post['name'], FILTER_VALIDATE_EMAIL)) {
-            $this->error['email'] = $this->language->get('error_name');
-        }
-
-        $team_info = $this->model_catalog_team->getTeamByEmail($this->request->post['email']);
-
-        if (!isset($this->request->get['team_id'])) {
-            if ($team_info) {
-                $this->error['warning'] = $this->language->get('error_exists_email');
-            }
-        } else {
-            if ($team_info && ($this->request->get['team_id'] != $team_info['team_id'])) {
-                $this->error['warning'] = $this->language->get('error_exists_email');
-            }
-        }
-
         if ((utf8_strlen(trim($this->request->post['name'])) < 1) || (utf8_strlen(trim($this->request->post['name'])) > 32)) {
             $this->error['name'] = $this->language->get('error_name');
         }
 
-        if ($this->request->post['password'] || (!isset($this->request->get['team_id']))) {
-            if ((utf8_strlen(html_entity_decode($this->request->post['password'], ENT_QUOTES, 'UTF-8')) < 4) || (utf8_strlen(html_entity_decode($this->request->post['password'], ENT_QUOTES, 'UTF-8')) > 40)) {
-                $this->error['password'] = $this->language->get('error_password');
-            }
+        if ((utf8_strlen(trim($this->request->post['designation'])) < 1) || (utf8_strlen(trim($this->request->post['designation'])) > 32)) {
+            $this->error['designation'] = $this->language->get('error_designation');
+        }
 
-            if ($this->request->post['password'] != $this->request->post['confirm']) {
-                $this->error['confirm'] = $this->language->get('error_confirm');
-            }
+        if ($this->error && !isset($this->error['warning'])) {
+            $this->error['warning'] = $this->language->get('error_warning');
         }
 
         return !$this->error;
